requirements.txt
streamlit
faiss-cpu
numpy
sentence_transformers
flask
pytesseract
pillow